delelte 메서드 응답 함수 수정

+++ post_app/views.py
@@ -55,5 +55,6 @@ def posts_detail(request, slug):
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
   # 요청 방식이 'DELETE'이면 찾아놓은 모델데이터를 DB 테이블에서 제거, 제거응답 상태 전송
-  post.delete()
-  return  Response(status=status.HTTP_204_NO_CONTENT)
+  elif request.method =='DELETE':
+    post.delete()
+    return  Response(status=status.HTTP_204_NO_CONTENT)
